+++ app/server/controller/menu_controller.go
@@ -19,53 +19,6 @@ func NewMenuController(mu domain.MenuUsecase) domain.MenuController {
 	}
 }
 
-type createMenuRequest struct {
-	OfferedAt                string `json:"offered_at" validate:"required,YYYY-MM-DD"`
-	PhotoUrl                 string `json:"photo_url" validate:"omitempty,url"`
-	ElementarySchoolCalories int32  `json:"elementary_school_calories" validate:"gt=0"`
-	JuniorHighSchoolCalories int32  `json:"junior_high_school_calories" validate:"gt=0"`
-	CityCode                 int32  `param:"code" validate:"required,gt=0"`
-}
-
-func (mc *menuController) Create(c echo.Context) error {
-	var req createMenuRequest
-
-	if err := c.Bind(&req); err != nil {
-
-		return c.JSON(errors.NewBadRequestError(err))
-	}
-
-	if err := c.Validate(&req); err != nil {
-		return c.JSON(errors.NewBadRequestError(err))
-	}
-
-	ctx := c.Request().Context()
-
-	offeredAt, err := util.ParseDate(req.OfferedAt)
-
-	if err != nil {
-		return c.JSON(errors.NewBadRequestError(err))
-	}
-
-	menu, err := domain.NewMenu(
-		offeredAt,
-		sql.NullString{String: req.PhotoUrl, Valid: req.PhotoUrl != ""},
-		req.ElementarySchoolCalories,
-		req.JuniorHighSchoolCalories,
-		req.CityCode,
-	)
-
-	if err != nil {
-		return c.JSON(errors.NewInternalServerError(err))
-	}
-
-	if err := mc.mu.Create(ctx, menu); err != nil {
-		return c.JSON(errors.NewInternalServerError(err))
-	}
-
-	return c.JSON(200, menu)
-}
-
 type getMenuRequest struct {
 	ID       string `param:"id" validate:"required,ulid"`
 	CityCode int32  `param:"code" validate:"required,gt=0"`
